Removed old GraphQL and replaced with single file without filters and netbox 4.3 compatible
