--- netbox_license/graphql/__init__.py
@@ -1,11 +1,0 @@
-from .schema import (
-    LicenseTypeQuery,
-    LicenseQuery,
-    LicenseAssignmentQuery,
-)
-
-schema = [
-    LicenseTypeQuery,
-    LicenseQuery,
-    LicenseAssignmentQuery,
-]
